fix: Don't crash when completing before starting to type.
If the user requests a completion but did not start typing an account
name yet, there's no account name in the AST under the cursor. This
possibility was previously unchecked and resulted in a nil-pointer
access. No we use the cursor position as the starting position of
the replaced text in the editor.